Slight change of wording in assignment 1, week 4
There was an inconsistency in the text.

The first case with four subsets of the data said:
> When calling graphlab.linear_regression.create(), use the same L2 penalty as
> before (i.e. l2_small_penalty).
The next case, using a larger penalty, said:
> With the argument `l2_penalty=1e5`, fit a 15th-order polynomial model on
> `set_1`, `set_2`, `set_3`, and `set_4`. Other than the extra parameter, the
> code should be the same as the experiment above.
There is no extra parameter -- both cases have the l2_penalty parameter. The
wording was changed to say "Other than the change in the l2_penalty value...".