Fix DeprecationWarning seen if using scikit-learn 0.17.0
visualize_document_clusters() produces warnings if using scikit-learn
0.17.0:
sklearn/utils/validation.py:386: DeprecationWarning: Passing 1d arrays as data is deprecated in 0.17 and willraise ValueError in 0.19. Reshape your data either using X.reshape(-1, 1) if your data has a single feature or X.reshape(1, -1) if it contains a single sample.

The problematic line is the call to pairwise_distances() because the
shape of `centroids[c]' is (547979,).

Per the suggestion in the warning message, use `centroids[c].reshape(1, -1)'
instead.